# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mix REST API",
    "description": "REST API for the Mix application - session management, messaging, and system operations",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/sessions/{id}/messages": {
      "get": {
        "description": "Retrieve all messages from a specific session",
        "operationId": "getSessionMessages",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageData"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of session messages"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session not found"
          }
        },
        "summary": "List session messages",
        "tags": [
          "Messages"
        ]
      },
      "post": {
        "description": "Send a user message to a specific session for AI processing",
        "operationId": "sendMessage",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "content": {
                    "description": "Message content to send",
                    "type": "string"
                  }
                },
                "required": [
                  "content"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageData"
                }
              }
            },
            "description": "Message sent and processed"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid message data"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session not found"
          }
        },
        "summary": "Send a message to session",
        "tags": [
          "Messages"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/RESTError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MessageData": {
        "properties": {
          "assistantResponse": {
            "description": "Assistant's response message (optional)",
            "type": "string"
          },
          "id": {
            "description": "Unique message identifier",
            "type": "string"
          },
          "reasoning": {
            "description": "Reasoning process (optional)",
            "type": "string"
          },
          "reasoningDuration": {
            "description": "Reasoning duration in milliseconds (optional)",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "sessionId": {
            "description": "Session identifier",
            "type": "string"
          },
          "toolCalls": {
            "description": "Tool calls made during message processing",
            "items": {
              "$ref": "#/components/schemas/ToolCallData"
            },
            "type": "array"
          },
          "userInput": {
            "description": "User's input message",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "role",
          "userInput"
        ],
        "type": "object"
      },
      "MessageRole": {
        "description": "Message role",
        "enum": [
          "user",
          "assistant",
          "tool"
        ],
        "type": "string"
      },
      "RESTError": {
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer"
          },
          "message": {
            "description": "Error message",
            "type": "string"
          },
          "type": {
            "description": "Error type",
            "enum": [
              "bad_request",
              "not_found",
              "internal_error",
              "unauthorized",
              "validation_error"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "type"
        ],
        "type": "object"
      },
...
      "ToolCallData": {
        "properties": {
          "finished": {
            "description": "Whether tool call has finished",
            "type": "boolean"
          },
          "id": {
            "description": "Unique tool call identifier",
            "type": "string"
          },
          "input": {
            "description": "Tool input parameters",
            "type": "string"
          },
          "isError": {
            "description": "Whether tool call resulted in error (optional)",
            "type": "boolean"
          },
          "name": {
            "description": "Tool name",
            "type": "string"
          },
          "result": {
            "description": "Tool execution result (optional)",
            "type": "string"
          },
          "type": {
            "description": "Tool type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "input",
          "type",
          "finished"
        ],
        "type": "object"
      }
    }
  }
}